% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\docType{data}
\name{items_df}
\alias{items_df}
\title{A tidy format of all current items for Valheim as collected from Jotunn: The Valheim Library}
\format{
a tibble with 434 rows and 5 variables:
\describe{
\item{prefab}{Base item in the game. Prefabs cannot be broken down into smaller constituent items.}
\item{token}{Reference token for a given item.}
\item{name}{In-game name of the item.}
\item{type}{The given type of an item.}
\item{description}{In game flavor text, where provided.}
}
}
\source{
{Jotunn: The Valheim Library, item list} \url{https://valheim-modding.github.io/Jotunn/data/objects/item-list.html}
}
\usage{
items_df
}
\description{
Token identifiers ($item_, $customization_) have been removed and variable
names converted to lowercase.
}
\details{
Items fall into one of the following types: Ammo, Bow, Chest, Consumable,
Customization, Helmet, Legs, Material, Misc, One Handed Weapon, Shield, Shoulder,
Tool, Torch, Trophie, Two Handed Weapon, or Utility.
}
\keyword{datasets}
