% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_raw.R
\docType{data}
\name{items_raw}
\alias{items_raw}
\title{All current items for Valheim, as collected from Jotunn: The Valheim Library}
\format{
a tibble with 434 rows and 5 variables:
\describe{
\item{Prefab}{Base item in the game. Prefabs cannot be broken down into smaller constituent items.}
\item{Token}{Reference token for a given item.}
\item{Name}{In-game name of the item.}
\item{Type}{The given type of an item.}
\item{Description}{In game flavor text, where provided.}
}
}
\source{
{Jotunn: The Valheim Library, item list} \url{https://valheim-modding.github.io/Jotunn/data/objects/item-list.html}
}
\usage{
items_raw
}
\description{
Items fall into one of the following types: Ammo, Bow, Chest, Consumable,
Customization, Helmet, Legs, Material, Misc, One Handed Weapon, Shield, Shoulder,
Tool, Torch, Trophie, Two Handed Weapon, or Utility.
}
\keyword{datasets}
