% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\docType{data}
\name{recipes_df}
\alias{recipes_df}
\title{A tidy format of all current recipes and required resources for Valheim,
as collected from Jotunn: The Valheim Library}
\format{
a tibble with 368 rows and 6 variables:
\describe{
\item{recipe_name}{Internal name of the recipe.}
\item{item_name}{In-game name of the item created by a recipe.}
\item{amount_created}{The amount of an item created by a recipe.}
\item{item_index}{A chronological numbering system for each item within a given recipe.}
\item{amount_required}{The amount of a given material required to make an item.}
\item{component_required}{The name of the item needed for a given recipe.}
}
}
\source{
{Jotunn: The Valheim Library, recipe list} \url{https://valheim-modding.github.io/Jotunn/data/objects/recipe-list.html}
}
\usage{
recipes_df
}
\description{
Variable names have been converted to snake_case and modified to be more
descriptive. Amount and component required were split into two separate
columns for ease of downstream use.
}
\keyword{datasets}
