% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_raw.R
\docType{data}
\name{recipes_raw}
\alias{recipes_raw}
\title{All current recipes and required resources for Valheim, as collected from
Jotunn: The Valheim Library}
\format{
a tibble with 141 rows and 4 variables:
\describe{
\item{Name}{Internal name of the recipe.}
\item{\verb{Item name}}{In-game name of the item created by a recipe.}
\item{Amount}{The amount of an item created by a recipe.}
\item{\verb{Resources required}}{The amount and type of each material required to make an item.}
}
}
\source{
{Jotunn: The Valheim Library, recipe list} \url{https://valheim-modding.github.io/Jotunn/data/objects/recipe-list.html}
}
\usage{
recipes_raw
}
\description{
All current recipes and required resources for Valheim, as collected from
Jotunn: The Valheim Library
}
\keyword{datasets}
